chore: remove middlewares module, moving decorator functionality to base and sqs
