+++ aws_lambda_powertools/utilities/batch/__init__.py
@@ -4,8 +4,7 @@
 Batch processing utility
 """
 
-from .base import BasePartialProcessor
-from .middlewares import batch_processor, sqs_batch_processor
-from .sqs import PartialSQSProcessor
+from .base import BasePartialProcessor, batch_processor
+from .sqs import PartialSQSProcessor, sqs_batch_processor
 
 __all__ = ("BasePartialProcessor", "PartialSQSProcessor", "batch_processor", "sqs_batch_processor")
